+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
 + Resolve [Issue 43](https://github.com/F5Networks/f5-declarative-onboarding/issues/43)
 + Resolve [Issue 28](https://github.com/F5Networks/f5-declarative-onboarding/issues/28)
 + Add support for Management Route
++ Add support for Route Domain [Issue 10](https://github.com/F5Networks/f5-declarative-onboarding/issues/10)
 
 ## Version 1.5.1
 + When running on BIG-IQ, poll TCW longer to match the TCW timeout